Move file-modifiers to dispatch-file-modifiers
