+++ biosimulations/libs/dispatch/api-models/src/lib/common/index.ts
@@ -1,3 +1,2 @@
 export * from './dispatch-simulation.model';
 export * from './dispatch-interfaces';
-export * from './file-modifier';
