+++ biosimulations/libs/dispatch/file-modifiers/.eslintrc
@@ -1,0 +1,1 @@
+{ "extends": "../../../.eslintrc", "rules": {}, "ignorePatterns": ["!**/*"] }
