+++ biosimulations/libs/dispatch/file-modifiers/README.md
@@ -1,0 +1,7 @@
+# dispatch-file-modifiers
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `ng test dispatch-file-modifiers` to execute the unit tests via [Jest](https://jestjs.io).
