+++ biosimulations/libs/dispatch/file-modifiers/jest.config.js
@@ -1,0 +1,14 @@
+module.exports = {
+  name: 'dispatch-file-modifiers',
+  preset: '../../../jest.config.js',
+  globals: {
+    'ts-jest': {
+      tsConfig: '<rootDir>/tsconfig.spec.json',
+    },
+  },
+  transform: {
+    '^.+\\.[tj]sx?$': 'ts-jest',
+  },
+  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'html'],
+  coverageDirectory: '../../../coverage/libs/dispatch/file-modifiers',
+};
