+++ biosimulations/tsconfig.base.json
@@ -60,6 +60,9 @@
       ],
       "@biosimulations/ontology/ontologies": [
         "libs/ontology/ontologies/src/index.ts"
+      ],
+      "@biosimulations/dispatch/file-modifiers": [
+        "libs/dispatch/file-modifiers/src/index.ts"
       ]
     }
   },
